refactor: replace `HttpContext` with `EmulationProvider` for clarity and accuracy
