+++ src/client/emulation/device/chrome.rs
