+++ src/client/emulation/device/firefox.rs
