+++ src/client/emulation/device/macros.rs
