+++ src/client/emulation/device/safari.rs
